+++ cli-tools/cli/bin/aofl-source.js
@@ -4,7 +4,8 @@ const program = require('commander');
 
 program
 .option('--repo [repo]', 'repo url')
+.option('-l, --list', 'List sourced modules')
 .parse(process.argv);
 
-const sourceModule = new SourceModule(program.args, program.repo);
+const sourceModule = new SourceModule(program.args, program.repo, program.list);
 sourceModule.init();